+++ packages/template-retail-react-app/app/components/toggle-card/index.jsx
@@ -49,7 +49,7 @@ export const ToggleCard = ({
                         <Heading
                             fontSize="lg"
                             lineHeight="30px"
-                            color={disabled && !editing && 'gray.400'}
+                            color={disabled && !editing && 'gray.600'}
                         >
                             {title}
                         </Heading>