+++ packages/commerce-sdk-react/CHANGELOG.md
@@ -1,5 +1,9 @@
-## v1.1.0-dev (Sep 26, 2023)
+## v1.2.0-dev (Nov 03, 2023)
+## v1.1.0 (Nov 03, 2023)
+
+- Add StorefrontPreview component [#1508](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1508)
 - Using `login` for authentication instead of `email` in registration process [#1464](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1464)
+- Fix bug to remove cookies using the default attributes used when setting cookie [#1505](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1505)
 
 ## v1.0.2 (Sep 21, 2023)